parse: Read raw data from file before processing
Read all available information from the files. Instead of merging the
results of different protocols, only merge multiple runs of the same
protocol at the same site (client/server).

Save the raw, parsed data so that the analysis can be re-run without
having to parse the data again.

The analysis combines the data as needed for each analysis step.